\name{predicts-functions-package}
\alias{predicts-functions-package}
\alias{predicts-functions}
\docType{package}
\title{
\packageTitle{predicts-functions}
}
\description{
\packageDescription{predicts-functions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{predicts-functions}
\packageIndices{predicts-functions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{predicts-functions}

Maintainer: \packageMaintainer{predicts-functions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
